Update image_time to run on Windows10 (and theoretically on any OS)
There is a function in the 'exifr' package that enables to extract the exif in R without calling an external function. It enables to run discr without installing Ruby + libexif + exif (which is currently not trivial in Windows 10). It only needs the packages 'exifr' and 'exifTool'. I haven't tested it on Mac OS yet, I should probably do so before updating the package.